Remove MainController and main.vtl and add get in AdvertController
