+++ src/main/java/com/codeclan/controllers/AdvertController.java
@@ -1,4 +1,48 @@
 package com.codeclan.controllers;
 
+import com.codeclan.db.DBHelper;
+import com.codeclan.db.Seeds;
+import com.codeclan.models.Advert;
+import spark.ModelAndView;
+import spark.template.velocity.VelocityTemplateEngine;
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import static spark.Spark.get;
+import static spark.SparkBase.staticFileLocation;
+
 public class AdvertController {
+
+    public static void main(String[] args) {
+        Seeds.seedData();
+
+        staticFileLocation("/public");
+
+        CategoryController categoryController = new CategoryController();
+        UserController userController = new UserController();
+        LoginController loginController = new LoginController();
+
+        get("/", (req, res) -> {
+
+            List<Advert> adverts = DBHelper.getAll(Advert.class);
+            Map<String, Object> model = new HashMap<>();
+            String loggedInUser = LoginController.getLoggedInUserName(req, res);
+            model.put("user", loggedInUser);
+            model.put("adverts", adverts);
+            model.put("template", "templates/adverts/index.vtl");
+
+            return new ModelAndView(model, "templates/layout.vtl");
+
+        }, new VelocityTemplateEngine());
+
+
+
+    }
+
+
+
+
+
 }
