+++ src/main/resources/templates/adverts/index.vtl
@@ -15,6 +15,6 @@
             <td>$advert.getAskingPrice()</td>
             <td>$advert.returnSimpleDateFormat()</td>
         </tr>
-            #end
+        #end
     </table>
 </div>
